+++ CHANGELOG.md
@@ -7,6 +7,11 @@ All notable changes to this project will be documented in this file.
 
 * `bundle` property on `ViewModelMapping`, that exposes recommended bundle to be used when searching for resources of given mapping.
 
+### Removed
+
+* Deprecated `MemoryStorageError.BatchInsertionReason` enum.
+* Deprecated `ViewModelMappingCustomizing` protocol.
+
 ## [7.4.0](https://github.com/DenTelezhkin/DTModelStorage/releases/tag/7.4.0)
 
 ### Added